\name{apply_prob_dtt}
\alias{apply_prob_dtt}

\title{
  Apply prob_dtt to output from shift.estimtes.
}
\description{
  Applies a set of birth-death models to a phylogeny.
}
\usage{
apply_prob_dtt(phylo, data, sampling.fractions, shift.res,
               combi = 1, backbone.option = "backbone2"){
}

\arguments{
  \item{phylo}{
an object of type 'phylo' (see ape documentation)
}
  ...
}

\details{
  
}

\value{
  A list of results from prob_dtt for subclades and backbone(s).
}
\references{
  Mazet, N., Morlon, H., Fabre, P-H. Condamine, F. (in prep)
}

\author{Nathan Mazet}

\examples{

}