\name{Caprimulgidae}
\alias{Caprimulgidae}
\docType{data}
\title{
The \emph{Caprimulgidae} phylogeny.
}
\description{
The MCC phylogeny for the \emph{Caprimulgidae}, from Jetz et al. (2012).
}
\usage{data("Caprimulgidae")}
\source{
Jetz, W., G. Thomas, J. Joy, K. Hartmann, and A. Mooers. 2012. The global diversity of birds in
space and time. \emph{Nature 491:444.}
}
\examples{
data("Caprimulgidae")

plot(Caprimulgidae)

}