\name{Cetacea_clades}
\alias{Cetacea_clades}
\docType{data}
\title{Stochastic map of clade membership in Cetacean phylogeny}
\description{simmap object of clade membership in Cetacean phylogeny}
\usage{data(Cetacea_clades)}
\details{
See \code{\link{Cetacea}}
}

\source{
Steeman ME et al.(2009) Radiation of extant cetaceans driven by restructuring of the oceans, \emph{Syst Biol} 58:573-585
}

\references{
Steeman ME et al.(2009) Radiation of extant cetaceans driven by restructuring of the oceans \emph{Syst Biol} 58:573-585

Morlon, H., Parsons, T.L., Plotkin, J.B. (2011) Reconciling molecular phylogenies with the fossil record \emph{Proc Nat Acad Sci} 108: 16327-16332

Condamine, F.L., Rolland, J., Morlon, H. (2013) Macroevolutionary perspectives to environmental change \emph{Eco Lett} 16: 72-85
}
\examples{
data(Cetacea_clades)
print(Cetacea_clades)
plot(Cetacea_clades)
}
\keyword{datasets}
