% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Calculate the cosine distance}
\usage{
cosine_distance(data)
}
\arguments{
\item{data}{a matrix of samples (sample x feature).}
}
\value{
a matrix of distances between samples (sample x sample).
}
\description{
Calculate the cosine distance
}
