% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Calculate the cosine similarity}
\usage{
cosine_similarity(data_1, data_2)
}
\arguments{
\item{data_1}{a matrix of the first set of samples (sample x feature).}

\item{data_2}{a matrix of the second set of samples (sample x feature).}
}
\value{
a matrix of similarity between two sets of samples.
}
\description{
Calculate the cosine similarity
}
