% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{dfPPI}
\alias{dfPPI}
\title{An example of protein-protein interaction data}
\usage{
data(dfPPI)
}
\description{
This is an example of protein-protein interaction data.
}
\details{
A two-column data frame consisting of binary interactions where each row represents an undirected edge (interaction) between two nodes (proteins) from two columns.
}
\keyword{datasets}

