% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_markers}
\alias{find_markers}
\title{Find marker genes based on Seurat}
\usage{
find_markers(counts, labels, n_markers = NULL)
}
\arguments{
\item{counts}{a matrix indicating the raw count expression of cells (gene x cell).}

\item{labels}{a vector indicating the corresponding cell type labels.}

\item{n_markers}{an integer indicating the number of marker genes to select for each cell type. The default is \code{NULL} and all marker genes will be output.}
}
\value{
a list containing marker genes for each cell type.
}
\description{
Find marker genes based on Seurat
}
