% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{GENE2GOtopLite}
\alias{GENE2GOtopLite}
\title{An example of GO annotation data for proteins}
\usage{
data(GENE2GOtopLite)
}
\description{
This is only an example of GO annotations for proteins. 
}
\details{
A list that maps GO ID to genes. The names should be gene symbols and the elements should be GO IDs. The list have the following format:

$SHC1

[1] "GO:0005158" "GO:0005068" "GO:0005159" "GO:0070435"

$POU5F1

[1] "GO:0035413" "GO:0003130" "GO:0060391" "GO:0090308" "GO:0060965" "GO:0035198"

$FGF12

[1] "GO:0008201"

......
}
\keyword{datasets}

