% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{GENE2KEGG}
\alias{GENE2KEGG}
\title{An example of KEGG annotation data for proteins}
\usage{
data(GENE2KEGG)
}
\description{
This is only an example of KEGG pathway annotations for proteins. 
}
\details{
A list that maps KEGG pathway ID to genes. The names should be gene symbols and the elements should be KEGG pathways. The list have the following format:

$IGSF5

[1] "hsa04530" "hsa05120"

$OR2T8

[1] "hsa04740"

$hCG_1776980

[1] "hsa00020" "hsa00190" "hsa01100" "hsa05010" "hsa05012" "hsa05016"

......
}
\keyword{datasets}

