\name{InfTemp}
\alias{InfTemp}
\docType{data}
\title{Inferred Temperature}
\description{Paleotemperature data}
\usage{data(InfTemp)}
\format{
  Dataframe of paleotemperature with 17632 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{a numeric vector - The age}
    \item{\code{Temperature}}{a numeric vector- The corresponding temperature}
  }
}
\examples{
data(InfTemp)
plot(InfTemp)
}
\keyword{datasets}
