% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_min_sampling}
\alias{max_min_sampling}
\title{Max-min sampling}
\usage{
max_min_sampling(data, n_points = 5)
}
\arguments{
\item{data}{a matrix of samples (sample x feature).}

\item{n_points}{an integer indicating the number of samples to take.}
}
\value{
a vector of the indices of the samples that have been sampled.
}
\description{
Max-min sampling
}
