\name{PANDA-package}
\alias{PANDA-package}
\alias{PANDA}
\docType{package}
\title{
PANDA
}
\description{
Implements fits of diversification models to phylogenetic data
}
\details{
\tabular{ll}{
Package: \tab PANDA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-27\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Hélène Morlon <morlon@ens.biologie.fr>
}
\references{
Morlon, H., Potts, M.D., Plotkin, J.B. (2010) Inferring the Dynamics of Diversification: a Coalescent Approach, PLoS B 8(9): e1000493

Morlon, H., Kemps, B., Plotkin, J.B., Brisson, D. (2012) Explosive radiation of a bacterial species group, Evolution, 66: 2577-2586

Morlon, H., Parsons, T.L. and Plotkin, J.B. (2011) Reconciling molecular phylogenies with the fossil record, Proc Nat Acad Sci 108: 16327-16332

Condamine, F.L., Rolland, J., and Morlon, H. (2013) Macroevolutionary perspectives to environmental change, Eco Lett 16: 72-85

Morlon, H. (2014) Phylogenetic approaches for studying diversification, Eco Lett
}
%\keyword{ phylogeny }
\seealso{
\code{geiger, laser, DDD, TreePar}
}

