% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panda.R
\name{panda}
\alias{panda}
\title{Make a panda}
\usage{
panda(msg = NULL, descriptor = NULL, panda = "random", stamp = TRUE)
}
\arguments{
\item{msg}{The panda can speak your words.}

\item{descriptor}{Select descriptor from [pandas].}

\item{panda}{Set a random seed to reproduce a panda.}

\item{stamp}{Set to `FALSE` if you wish to remove details text in bottom
right of image.}
}
\description{
Make a panda
}
