% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_archetypes}
\alias{plot_archetypes}
\title{Visualization of archetypes}
\usage{
plot_archetypes(
  Z,
  phi,
  f_name = "archetypes",
  save_dir = ".",
  width = 7,
  height = 7
)
}
\arguments{
\item{Z}{a matrix indicating the normalized expression of cells (cell x gene).}

\item{phi}{a matrix indicating the expression of archetypes (archetype x gene).}

\item{f_name}{a character string naming the pdf file of the figure. The default is "archetypes".}

\item{save_dir}{a character string indicating the directory where the pdf file is saved. The default is the current directory.}

\item{width}{a value indicating the width of the figure. The default is 7.}

\item{height}{a value indicating the height of the figure. The default is 7.}
}
\value{
a pdf file containing the figure.
}
\description{
Visualization of archetypes
}
