\name{RPANDA-package}
\alias{RPANDA-package}
\alias{RPANDA}
\docType{package}
\title{
RPANDA
}
\description{
The package helps to fit various diversity models
}
\details{
\tabular{ll}{
Package: \tab RPANDA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-27\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Hélène Morlon

Maintainer: Hélène Morlon <helene.morlon@cmap.polytechnique.fr>
}
\references{
Hélène Morlon, Matthew D. Potts, Joshua B. Plotkin: \emph{Inferring the Dynamics of Diversification: a Coalescent Approach}, PLoS Biology September 2010, Volume 8, Issue 9

Hélène Morlon, Todd L. Parsons, Joshua B. Plotkin:  \emph{Reconciling molecular phylogenies with the fossil record}, PNAS 2011

Fabien L. Condamine, Jonathan Rolland, Hélène Morlon: \emph{Macroevolutionary perspectives to environmental change}, Ecology Letters (2013)
}
%\keyword{ philogeny }
\seealso{
\code{geiger, laser, DDD, TreePar}
}

