% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvr.R
\docType{data}
\name{rvr}
\alias{rvr}
\title{Reproducibility v replicability, with pandas}
\format{An object of class \code{patchwork} (inherits from \code{gg}, \code{ggplot}) of length 10.}
\usage{
rvr
}
\description{
This is the motivating plot for the [panda] package. It'll go in thesis
somewhere.
}
\details{
See README for the generation of this patchwork plot.
}
\keyword{datasets}
