% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANDA.R
\name{sc_initialize}
\alias{sc_initialize}
\title{Initialize the combination coefficients for cells and archetypes}
\usage{
sc_initialize(Z, n_archetypes)
}
\arguments{
\item{Z}{a matrix indicating the normalized expression of cells (cell x gene).}

\item{n_archetypes}{an integer indicating the number of archetypes.}
}
\value{
a list containing the following components:
\describe{
  \item{\code{alpha}}{a matrix of the initial combination coefficients for cells (cell x archetype).}
  \item{\code{W}}{a matrix of the initial combination coefficients for archetypes (archetype x cell).}
}
}
\description{
Initialize the combination coefficients for cells and archetypes
}
