% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANDA.R
\name{sc_model}
\alias{sc_model}
\title{Find archetypes for one cell type}
\usage{
sc_model(
  X,
  Z,
  N,
  n_archetypes,
  tol = 1e-08,
  maxiter = 2000,
  do_initialize = TRUE,
  do_plot = TRUE,
  f_name = "archetypes",
  save_dir = "."
)
}
\arguments{
\item{X}{a matrix indicating the raw count expression of cells (cell x gene).}

\item{Z}{a matrix indicating the normalized expression of cells (cell x gene).}

\item{N}{a vector indicating the library size of each cell.}

\item{n_archetypes}{an integer indicating the number of archetypes.}

\item{tol}{a value indicating the convergence threshold. The default is 1e-8.}

\item{maxiter}{an integer indicating the maximum number of iterations. The default is 2000.}

\item{do_initialize}{a logical value indicating whether to perform initialization. The default is \code{TRUE}.}

\item{do_plot}{a logical value indicating whether to visualize the archetypes using UMAP (Uniform Manifold Approximation and Projection). The default is \code{TRUE}.}

\item{f_name}{a character string naming the pdf file of the UMAP. The default is "archetypes".}

\item{save_dir}{a character string indicating the directory where the pdf file is saved. The default is the current directory.}
}
\value{
a list containing the following components:
\describe{
  \item{\code{phi}}{a matrix of the expression of archetypes (archetype x gene).}
  \item{\code{alpha}}{a matrix of combination coefficients for cells (cell x archetype).}
  \item{\code{W}}{a matrix of combination coefficients for archetypes (archetype x cell).}
}
}
\description{
Find archetypes for one cell type
}
