% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANDA.R
\name{sc_preprocess}
\alias{sc_preprocess}
\title{Preprocess the scRNA-seq reference}
\usage{
sc_preprocess(
  sc_counts,
  sc_labels,
  min_counts = 200,
  do_hvgs = TRUE,
  do_markers = TRUE,
  n_hvgs = 2000,
  n_markers = 20,
  n_sample_cells = 500
)
}
\arguments{
\item{sc_counts}{a matrix indicating the raw count expression of cells (cell x gene).}

\item{sc_labels}{a vector indicating the corresponding cell type labels.}

\item{min_counts}{an integer indicating the minimum total counts for cells to be included. The default is 200.}

\item{do_hvgs}{a logical value indicating whether to find highly variable genes. The default is \code{TRUE}.}

\item{do_markers}{a logical value indicating whether to find marker genes for each cell type. The default is \code{TRUE}.}

\item{n_hvgs}{an integer indicating the number of highly variable genes to select. The default is 2000.}

\item{n_markers}{an integer indicating the number of marker genes to select for each cell type. The default is 20.}

\item{n_sample_cells}{an integer indicating the maximum number of cells contained in each cell type. Cell types with a cell count exceeding \code{n_sample_cells} will be downsampled to \code{n_sample_cells}. The default is 500.}
}
\value{
a list containing the following components:
\describe{
  \item{\code{X_list}}{a list containing the preprocessed count matrix for each cell type.}
  \item{\code{Z_list}}{a list containing the preprocessed normalized expression matrix for each cell type.}
  \item{\code{Z_all_list}}{a list containing the preprocessed normalized expression matrix with all genes for each cell type.}
  \item{\code{N_list}}{a list containing the library size vector for each cell type.}
  \item{\code{unique_labels}}{a vector of unique cell types included in the scRNA-seq reference.}
}
}
\description{
Preprocess the scRNA-seq reference
}
