% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANDA.R
\name{sc_train}
\alias{sc_train}
\title{Archetypal analysis on scRNA-seq reference}
\usage{
sc_train(
  sc_counts,
  sc_labels,
  n_archetypes_vec = 10,
  tol = 1e-08,
  maxiter = 2000,
  do_initialize = TRUE,
  do_plot = TRUE,
  min_counts = 200,
  do_hvgs = TRUE,
  do_markers = TRUE,
  n_hvgs = 2000,
  n_markers = 20,
  n_sample_cells = 500,
  do_parallel = TRUE,
  n_cores = 20,
  save_res = TRUE,
  save_dir = "."
)
}
\arguments{
\item{sc_counts}{a matrix indicating the raw count expression of cells (cell x gene).}

\item{sc_labels}{a vector indicating the corresponding cell type labels.}

\item{n_archetypes_vec}{either an integer or a vector indicating the number of archetypes for each cell type. If it is an integer, the number of archetypes is the same for all cell types and is determined by this integer. If specifying a different number of archetypes for different cell types, set \code{n_archetypes_vec} to a vector with a length equal to the number of cell types and name each element according to the respective cell type. The default is 10.}

\item{tol}{a value indicating the convergence threshold. The default is 1e-8.}

\item{maxiter}{an integer indicating the maximum number of iterations. The default is 2000.}

\item{do_initialize}{a logical value indicating whether to perform initialization. The default is \code{TRUE}.}

\item{do_plot}{a logical value indicating whether to visualize the archetypes using UMAP (Uniform Manifold Approximation and Projection). The default is \code{TRUE}.}

\item{min_counts}{an integer indicating the minimum total counts for cells to be included. The default is 200.}

\item{do_hvgs}{a logical value indicating whether to find highly variable genes. The default is \code{TRUE}.}

\item{do_markers}{a logical value indicating whether to find marker genes for each cell type. The default is \code{TRUE}.}

\item{n_hvgs}{an integer indicating the number of highly variable genes used for archetypal analysis. The default is 2000.}

\item{n_markers}{an integer indicating the number of marker genes used for archetypal analysis for each cell type. The default is 20.}

\item{n_sample_cells}{an integer indicating the maximum number of cells contained in each cell type. Cell types with a cell count exceeding \code{n_sample_cells} will be downsampled to \code{n_sample_cells}. The default is 500.}

\item{do_parallel}{a logical value indicating whether to use parallel computation. The default is \code{TRUE}.}

\item{n_cores}{an integer indicating the number of cores used for parallel computation. The default is 20.}

\item{save_res}{a logical value indicating whether to save the results. The default is \code{TRUE}.}

\item{save_dir}{a character string indicating the directory where the results are saved. The default is the current directory.}
}
\value{
a list containing the following components:
\describe{
  \item{\code{archetypes_all_list}}{a list containing the archetypes with all genes for each cell type.}
  \item{\code{archetypes_list}}{a list containing the archetypes with selected genes (highly variable genes and marker genes) for each cell type.}
  \item{\code{res_list}}{a list containing all results of archetypal analysis for each cell type.}
  \item{\code{run_time}}{a value of the running time for archetypal analysis.}
  \item{\code{peak_ram}}{a value of the peak memory usage for archetypal analysis.}
}
}
\description{
Archetypal analysis on scRNA-seq reference
}
