\name{shifts_cetacea}
\alias{shifts_cetacea}
\docType{data}
\title{Cetacean shift.estimates results}
\description{Results of shift.estimates applyied to Cetaceans}
\usage{data(shifts_cetacea)}
\details{
This object is the result of shifts.estimates applied to the Cetacean phylogeny as in the example of shift.estimates function.
}

\source{
Mazet et al. (2022) Estimating clade-specific diversification rates and palaeodiversity dynamics from reconstructed phylogenies, \emph{BioRxiv}
}

\references{
Steeman ME et al.(2009) Radiation of extant cetaceans driven by restructuring of the oceans \emph{Syst Biol} 58:573-585

Morlon, H., Parsons, T.L., Plotkin, J.B. (2011) Reconciling molecular phylogenies with the fossil record \emph{Proc Nat Acad Sci} 108: 16327-16332

Mazet et al. (2022) Estimating clade-specific diversification rates and palaeodiversity dynamics from reconstructed phylogenies, \emph{BioRxiv}
}
\examples{
data(shifts_cetacea)
print(shifts_cetacea)
}
\keyword{datasets}
