\name{simudat}
\alias{simudat}
\docType{data}
\title{
Example data set for PanDM
}
\description{
\verb{simudat} is the manually generated simulation data for PanDM function. \verb{simudat} is composed of the \emph{p-values} of 10,000 CpG sites from initial separate DM calling for 5 cancer types.
}
\usage{data("simudat")}
\format{

}
\details{

}
\source{

}
\references{
Shi, M., Tsui, S., Wu, H., Wei, Y.,(2017) \verb{Pan-cancer analysis of differential DNA methylation patterns}
}
\examples{
data(simudat)
dim(simudat)
head(simudat)
## maybe str(simudat) ; plot(simudat) ...
}
\keyword{datasets}
