% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANDA.R
\name{st_model}
\alias{st_model}
\title{The core function for deconvolution}
\usage{
st_model(Y, N, phi, Q, sigma = 0.3, tol = 1e-08, maxiter = 2000)
}
\arguments{
\item{Y}{a matrix indicating the raw count expression of spots (spot x gene).}

\item{N}{a vector indicating the total counts of each spot.}

\item{phi}{a matrix indicating the expression of archetypes (archetype x gene)}

\item{Q}{a one-hot matrix indicating the cell type corresponding to each archetype (cell type x archetype).}

\item{sigma}{a value indicating the parameter that determine the standard deviation of the normal distribution for the platform effect. The default is 0.3.}

\item{tol}{a value indicating the convergence threshold. The default is 1e-8.}

\item{maxiter}{an integer indicating the maximum number of iterations. The default is 2000.}
}
\value{
a list containing the following components:
\describe{
  \item{\code{proportion}}{a matrix of the cell type proportions for spots (spot x cell type).}
  \item{\code{beta}}{a matrix of the cell type abundance for spots (spot x cell type).}
  \item{\code{alpha}}{a matrix of the combination coefficients for spots (spot x archetype).}
  \item{\code{omega}}{a matrix of the auxiliary matrix (spot x archetype).}
  \item{\code{gamma}}{a vector of the platform effect for genes.}
}
}
\description{
The core function for deconvolution
}
