% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PANDA.R
\name{st_preprocess}
\alias{st_preprocess}
\title{Preprocess the spatial transcriptomics data}
\usage{
st_preprocess(st_counts, n_hvgs = 5000, keep_genes = NULL)
}
\arguments{
\item{st_counts}{a matrix indicating the raw count expression of spots (spot x gene).}

\item{n_hvgs}{an integer indicating the number of highly variable genes to select. The default is 5000.}

\item{keep_genes}{a vector of genes from which the highly variable genes will be selected. If \code{NULL}, the highly variable genes will be selected from all genes in the \code{st_counts}.}
}
\value{
a list containing the following components:
\describe{
  \item{\code{Y}}{a matrix of the preprocessed count matrix (spot x gene).}
  \item{\code{N}}{a vector of total counts for spot.}
}
}
\description{
Preprocess the spatial transcriptomics data
}
