\name{taxo_cetacea}
\alias{taxo_cetacea}
\docType{data}
\title{Cetacean taxonomy}
\description{Taxonomy of Cetaceans}
\usage{data(taxo_cetacea)}
\details{
This taxonomy lists all species of Cetaceans to properly calculate sampling fractions by clades. It corresponds to the phylogeny of Steeman et al. (2009).
}

\source{
	Mazet, N., Morlon, H., Fabre, P-H. Condamine, F. (2023). Estimating clade-specific diversification rates and palaeodiversity dynamics from reconstructed phylogenies. BioRχiv.
}

\references{
Steeman ME et al.(2009) Radiation of extant cetaceans driven by restructuring of the oceans \emph{Syst Biol} 58:573-585

Morlon, H., Parsons, T.L., Plotkin, J.B. (2011) Reconciling molecular phylogenies with the fossil record \emph{Proc Nat Acad Sci} 108: 16327-16332

Mazet et al. (2022) Estimating clade-specific diversification rates and palaeodiversity dynamics from reconstructed phylogenies, \emph{BioRxiv}
}
\examples{
data(taxo_cetacea)
print(taxo_cetacea)
}
\keyword{datasets}
