\name{MPhiFFT}
\alias{MPhiFFT}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A class used internally to compute ClaDS's likelihood}
\description{
This class represents a matrix A = (1/rowSums(Toep)) * Toep where Toep is a Toeplitz matrix.}
\references{
Maliet O., Hartig F. and Morlon H. 2019, A model with many small shifts for estimating species-specific diversificaton rates, \emph{Nature Ecology and Evolution} in press
}
%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{fit_ClaDS}}
}

