% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panda.R
\name{panda}
\alias{panda}
\title{Make a panda}
\usage{
panda(msg = NULL, mood = NULL, stamp = TRUE, panda = "random")
}
\arguments{
\item{msg}{The panda can speak your words.}

\item{mood}{Select mood from [panda_mood].}

\item{stamp}{Set to `FALSE` if you wish to remove details text in bottom
right of image.}

\item{seed}{Set a random seed to reproduce a panda.}
}
\description{
Make a panda
}
