% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandas.R
\docType{data}
\name{pandas}
\alias{pandas}
\title{Preset pandas}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 3 columns.}
\usage{
pandas
}
\description{
Select [panda] by mood or name.
}
\keyword{datasets}
