\name{RPANDA-package}
\alias{RPANDA-package}
\alias{RPANDA}
\docType{package}
\title{
RPANDA
}
\description{
Implements eco-evolutionary analyses on phylogenetic trees
}
\details{
\tabular{ll}{
Package: \tab RPANDA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-27\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Hélène Morlon <morlon@biologie.ens.fr>

Fabien Condamine <fabien.condamine@gmail.com>

Eric Lewitus <lewitus@biologie.ens.fr>

Marc Manceau <manceau@biologie.ens.fr>
}
\references{
Morlon, H., Potts, M.D., Plotkin, J.B. (2010) Inferring the dynamics of diversification: a coalescent approach, PLoS B 8(9): e1000493

Morlon, H., Parsons, T.L. and Plotkin, J.B. (2011) Reconciling molecular phylogenies with the fossil record, Proc Nat Acad Sci 108: 16327-16332

Morlon, H., Kemps, B., Plotkin, J.B., Brisson, D. (2012) Explosive radiation of a bacterial species group, Evolution, 66: 2577-2586

Condamine, F.L., Rolland, J., and Morlon, H. (2013) Macroevolutionary perspectives to environmental change, Eco Lett 16: 72-85

Morlon, H. (2014) Phylogenetic approaches for studying diversification, Eco Lett 7: 508-525

Manceau, M., Lambert, A., Morlon, H. (2015) Phylogenies support out-of-equilibrium models of biodiversity Eco Lett 18: 347-356

Lewitus E., Morlon H. (submitted) A graph-theoretical approach to characterising and comparing phylogenies
}
%\keyword{ phylogeny }


